requirements.txt
# ── LexAssist AI — Python dependencies ───────────────────────────────────────
# Install: pip install -r requirements.txt
# Python 3.11+ required

# ── Core agentic framework ────────────────────────────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0

# ── Vector store & embeddings ─────────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit>=1.40.0

# ── Utilities ─────────────────────────────────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0.0
groq>=0.9.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0
